@@ -2,23 +2,23 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>netcoreapp2.1;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+    <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>liquid</ToolCommandName>
     <PackageOutputPath>./nupkg</PackageOutputPath>
     <RepositoryUrl>https://github.com/kurtmkurtm/LiquidTestReports</RepositoryUrl>
     <PackageProjectUrl>https://github.com/kurtmkurtm/LiquidTestReports</PackageProjectUrl>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <PackageIcon>nuget_cli.png</PackageIcon>    
+    <PackageIcon>nuget_cli.png</PackageIcon>
     <Description>
       LiquidTestReports.Cli Generates Markdown and Custom Test Reports from TRX and or JUnit test files.
-      The output will look like this https://github.com/kurtmkurtm/LiquidTestReports/blob/master/docs/samples/cli/SingleInput.md
     </Description>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.FileSystemGlobbing" Version="5.0.0" />
-    <PackageReference Include="Spectre.Console" Version="0.42.0" />
-    <PackageReference Include="System.CommandLine.DragonFruit" Version="0.3.0-alpha.21216.1" />
+    <PackageReference Include="Microsoft.Extensions.FileSystemGlobbing" Version="8.0.0" />
+    <PackageReference Include="Spectre.Console" Version="0.48.0" />
+    <PackageReference Include="System.CommandLine.DragonFruit" Version="0.4.0-alpha.22272.1" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\LiquidTestReports.Core\LiquidTestReports.Core.csproj" />
@@ -29,6 +29,10 @@
       <PackagePath></PackagePath>
     </None>
     <None Include="..\..\LICENSE">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
+    <None Include="README.md">
       <Pack>True</Pack>
       <PackagePath></PackagePath>
     </None>
@@ -51,7 +55,8 @@
     <Authors>Kurt Murrell</Authors>
     <PackageId>LiquidTestReports.Cli</PackageId>
     <Product>LiquidTestReports.Cli</Product>
-    <Copyright>2021</Copyright>
-    <Description>.NET tool to combine and convert TRX and JUnit tests reports into Markdown with Liquid template support</Description>
+    <Copyright>2024</Copyright>
+    <Description>.NET tool to combine and convert TRX and JUnit tests reports into Markdown with
+      Liquid template support</Description>
   </PropertyGroup>
 </Project>
