@@ -1,20 +1,22 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;net461</TargetFrameworks>
+    <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="JunitXml.TestLogger" Version="3.0.87" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.1" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.2.3" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.2.3" />
+    <PackageReference Include="JunitXml.TestLogger" Version="3.1.12" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
+    <PackageReference Include="MSTest.TestAdapter" Version="3.2.1" />
+    <PackageReference Include="MSTest.TestFramework" Version="3.2.1" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\LiquidTestReports.Custom\LiquidTestReports.Custom.csproj" />
-    <ProjectReference Include="..\..\..\src\LiquidTestReports.Markdown\LiquidTestReports.Markdown.csproj" />
+    <ProjectReference
+      Include="..\..\..\src\LiquidTestReports.Custom\LiquidTestReports.Custom.csproj" />
+    <ProjectReference
+      Include="..\..\..\src\LiquidTestReports.Markdown\LiquidTestReports.Markdown.csproj" />
     <ProjectReference Include="..\SampleProject\SampleProject.csproj" />
   </ItemGroup>
 
