@@ -1,23 +1,25 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;net461</TargetFrameworks>
+    <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="JunitXml.TestLogger" Version="3.0.87" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.1" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+    <PackageReference Include="JunitXml.TestLogger" Version="3.1.12" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
+    <PackageReference Include="xunit" Version="2.7.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.5.7">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\LiquidTestReports.Custom\LiquidTestReports.Custom.csproj" />
-    <ProjectReference Include="..\..\..\src\LiquidTestReports.Markdown\LiquidTestReports.Markdown.csproj" />
+    <ProjectReference
+      Include="..\..\..\src\LiquidTestReports.Custom\LiquidTestReports.Custom.csproj" />
+    <ProjectReference
+      Include="..\..\..\src\LiquidTestReports.Markdown\LiquidTestReports.Markdown.csproj" />
     <ProjectReference Include="..\SampleProject\SampleProject.csproj" />
   </ItemGroup>
 
