@@ -3,6 +3,7 @@ using LiquidTestReports.Core.Models;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Xunit;
 
 namespace LiquidTestReports.Cli.Tests
@@ -22,16 +23,16 @@ namespace LiquidTestReports.Cli.Tests
             //Arrange
             var groupTitleTest = "groupTitleTest.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, groupTitleTest));
-            var files = new List<ReportInput>();
+            var files = new List<string>();
 
             // Group by title, add test framework as suffix
-            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*netcoreapp3.1-sample.trx"))
-                files.Add(new ReportInput($"File={file};GroupTitle=.NET Core 3.1;TestSuffix= ({file.Name.Split('-')[0]})"));
-            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*net461-sample.trx"))
-                files.Add(new ReportInput($"File={file};GroupTitle=.NET Framework 4.6.1;TestSuffix= ({file.Name.Split('-')[0]})"));
+            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*net8.0-sample.trx"))
+                files.Add($"File={file};GroupTitle=.NET 8.0;TestSuffix= ({file.Name.Split('-')[0]})");
+            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*net6.0-sample.trx"))
+                files.Add($"File={file};GroupTitle=.NET 6.0;TestSuffix= ({file.Name.Split('-')[0]})");
 
             // Act
-            Program.Main(files.ToArray(), destinationReport);
+            Program.Main(files, destinationReport);
 
             // Assert
             Assert.True(destinationReport.Exists);
@@ -43,16 +44,16 @@ namespace LiquidTestReports.Cli.Tests
             //Arrange
             var testSuffixTest = "testSuffixTest.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, testSuffixTest));
-            var files = new List<ReportInput>();
+            var files = new List<string>();
 
             // Group by test framework, add target framework as suffix
-            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*netcoreapp3.1-sample.trx"))
-                files.Add(new ReportInput($"File={file};TestSuffix= (.NET Core 3.1)"));
-            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*net461-sample.trx"))
-                files.Add(new ReportInput($"File={file};TestSuffix= (.NET Framework 4.6.1)"));
+            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*net8.0-sample.trx"))
+                files.Add($"File={file};TestSuffix= (.NET 8.0)");
+            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*net6.0-sample.trx"))
+                files.Add($"File={file};TestSuffix= (.NET 6.0)");
 
             // Act
-            Program.Main(files.ToArray(), destinationReport);
+            Program.Main(files, destinationReport);
 
             // Assert
             Assert.True(destinationReport.Exists);
@@ -64,14 +65,14 @@ namespace LiquidTestReports.Cli.Tests
             //Arrange
             var fileOnlyTest = "fileOnlyTest.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, fileOnlyTest));
-            var files = new List<ReportInput>();
+            var files = new List<string>();
 
             // No grouping or suffix
             foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*sample.trx"))
-                files.Add(new ReportInput($"File={file}"));
+                files.Add($"File={file}");
 
             // Act
-            Program.Main(files.ToArray(), destinationReport);
+            Program.Main(files, destinationReport);
 
             // Assert
             Assert.True(destinationReport.Exists);
@@ -84,14 +85,14 @@ namespace LiquidTestReports.Cli.Tests
             var title = "My Test Title";
             var titleTest = "titleTest.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, titleTest));
-            var files = new List<ReportInput>();
+            var files = new List<string>();
 
             // No grouping or suffix
             foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("*sample.trx"))
-                files.Add(new ReportInput($"File={file}"));
+                files.Add($"File={file}");
 
             // Act
-            Program.Main(files.ToArray(), destinationReport, title);
+            Program.Main(files, destinationReport, title);
             // Assert
             Assert.True(destinationReport.Exists);
         }
@@ -103,19 +104,23 @@ namespace LiquidTestReports.Cli.Tests
             var title = "My Full Stack Test Report (JUnit + TRX)";
             var titleTest = "junitTest.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, titleTest));
-            var files = new List<ReportInput>();
+            var files = new List<string>();
 
-            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputJUnitDirectory)).GetFiles("xUnit-netcoreapp3.1-junit-sample.xml"))
-                files.Add(new ReportInput($"File={file};Format=JUnit;GroupTitle=JUnit Tests"));
+            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputJUnitDirectory)).GetFiles("xUnit-net8.0-junit-sample.xml"))
+                files.Add($"File={file};Format=JUnit;GroupTitle=JUnit Tests");
 
-            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("xUnit-netcoreapp3.1-sample.trx"))
-                files.Add(new ReportInput($"File={file};Format=Trx;GroupTitle=Trx Tests"));
+            foreach (var file in new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, _inputTrxDirectory)).GetFiles("xUnit-net8.0-sample.trx"))
+                files.Add($"File={file};Format=Trx;GroupTitle=Trx Tests");
 
             // Act
-            Program.Main(files.ToArray(), destinationReport, title);
+            Program.Main(files, destinationReport, title);
 
             // Assert
             Assert.True(destinationReport.Exists);
+            var content = File.ReadAllText(destinationReport.FullName);
+            Assert.Contains("# My Full Stack Test Report (JUnit + TRX)", content);
+            Assert.Contains("#### JUnit Tests", content);
+            Assert.Contains("#### Trx Tests", content);
         }
 
         [Fact]
@@ -125,17 +130,20 @@ namespace LiquidTestReports.Cli.Tests
             var title = "My Full Stack Test Report (JUnit + TRX)";
             var titleTest = "junitTest.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, titleTest));
-            var files = new List<ReportInput>
+            var files = new[]
             {
-                new ReportInput($"File=**/*junit-sample.xml;Folder={Environment.CurrentDirectory};Format=JUnit;GroupTitle=JUnit Tests"),
-                new ReportInput($"File=**/*sample.trx;Folder={Environment.CurrentDirectory};Format=Trx;GroupTitle=Trx Tests")
+                $"File=**/*junit-sample.xml;Folder={Environment.CurrentDirectory};Format=JUnit;GroupTitle=JUnit Tests",
+                $"File=**/*sample.trx;Folder={Environment.CurrentDirectory};Format=Trx;GroupTitle=Trx Tests"
             };
 
             // Act
-            Program.Main(files.ToArray(), destinationReport, title);
+            Program.Main(files, destinationReport, title);
 
             // Assert
             Assert.True(destinationReport.Exists);
+            var content = File.ReadAllText(destinationReport.FullName);
+            Assert.Contains("#### JUnit Tests", content);
+            Assert.Contains("#### Trx Tests", content);
         }
 
 
@@ -146,13 +154,13 @@ namespace LiquidTestReports.Cli.Tests
             var title = "My Full Stack Test Report (TRX)";
             var titleTest = "globTrxJunit.md";
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, titleTest));
-            var files = new List<ReportInput>
+            var files = new[]
             {
-                new ReportInput($"File=*/*xunit*.trx;Format=Trx;GroupTitle=Trx Tests")
+                $"File=*/*xunit*.trx;Format=Trx;GroupTitle=Trx Tests"
             };
 
             // Act
-            Program.Main(files.ToArray(), destinationReport, title);
+            Program.Main(files, destinationReport, title);
 
             // Assert
             Assert.True(destinationReport.Exists);
@@ -169,12 +177,12 @@ namespace LiquidTestReports.Cli.Tests
             var destinationReport = new FileInfo(Path.Combine(_outputFolder, reportName));
             var input = new[]
             {
-                new ReportInput($"File=**/*sample.trx;Folder={Environment.CurrentDirectory};Format=Trx;RunId=123")
+                $"File=**/*sample.trx;Folder={Environment.CurrentDirectory};Format=Trx;RunId=123"
             };
 
             var templatePath = Path.Combine(Environment.CurrentDirectory, _inputTemplateDirectory, "example.md");
             var templateFileInfo = new FileInfo(templatePath);
-            var parameters = new ParametersInput("Environment=UAT;TicketId=abc123");
+            var parameters = "Environment=UAT;TicketId=abc123";
 
             // Act
             Program.Main(input, destinationReport, template: templateFileInfo, parameters: parameters);
